+++ CMakeLists.txt
@@ -120,10 +120,10 @@ file(GLOB srcs CONFIGURE_DEPENDS "src/*.cpp" "src/*.h" "src/*.inl")
 get_filename_component(full_path_main_cpp ${CMAKE_CURRENT_SOURCE_DIR}/src/main.cpp ABSOLUTE)
 list(REMOVE_ITEM srcs ${full_path_main_cpp})
 
-# Ensure settings.cpp is always in the library (persistence for native + Emscripten)
-set(settings_cpp ${CMAKE_CURRENT_SOURCE_DIR}/src/settings.cpp)
-list(REMOVE_ITEM srcs "src/settings.cpp" ${settings_cpp})
-list(APPEND srcs ${settings_cpp})
+# Ensure utl_settings.cpp is always in the library (persistence for native + Emscripten)
+set(utl_settings_cpp ${CMAKE_CURRENT_SOURCE_DIR}/src/utl_settings.cpp)
+list(REMOVE_ITEM srcs "src/utl_settings.cpp" ${utl_settings_cpp})
+list(APPEND srcs ${utl_settings_cpp})
 
 set(gui_settings_cpp ${CMAKE_CURRENT_SOURCE_DIR}/src/gui_settings.cpp)
 list(REMOVE_ITEM srcs "src/gui_settings.cpp" ${gui_settings_cpp})
