+++ src/gui_settings.cpp
@@ -10,7 +10,7 @@
 #include "gui.h"
 #include "imgui.h"
 #include "occt_view.h"
-#include "settings.h"
+#include "utl_settings.h"
 #include "sketch.h"
 #include "sketch_nodes.h"
