+++ src/utl_settings.cpp
@@ -1,4 +1,4 @@
-#include "settings.h"
+#include "utl_settings.h"
 
 #include <cassert>
 #include <cstdlib>
@@ -51,6 +51,52 @@ std::filesystem::path user_settings_json_path()
 #endif
 }
 
+std::string load_with_defaults()
+{
+  std::string content;
+#ifdef __EMSCRIPTEN__
+  void* ptr = (void*)(intptr_t)EM_ASM_INT({
+    var s   = localStorage.getItem("ezycad_settings") || "";
+    var len = lengthBytesUTF8(s) + 1;
+    var buf = _malloc(len);
+    if (buf)
+      stringToUTF8(s, buf, len);
+    return buf;
+  });
+  if (ptr)
+  {
+    content = (const char*)ptr;
+    free(ptr);
+  }
+#else
+  auto read_file = [](const std::filesystem::path& p) -> std::string
+  {
+    std::ifstream f(p, std::ios::binary);
+    if (!f)
+      return {};
+    std::ostringstream os;
+    os << f.rdbuf();
+    return os.str();
+  };
+
+  const std::filesystem::path user_p = user_settings_json_path();
+  if (!user_p.empty())
+    content = read_file(user_p);
+  if (content.empty())
+  {
+    // Legacy: cwd ezycad_settings.json (same directory as exe when launched that way).
+    content = read_file(std::filesystem::path("ezycad_settings.json"));
+  }
+#endif
+  if (content.empty())
+  {
+    content = load_defaults();
+    if (!content.empty())
+      save(content);
+  }
+  return content;
+}
+
 std::string load_defaults()
 {
   if (s_log_callback)
@@ -82,18 +128,6 @@ std::string load_defaults()
     s_log_callback("Settings: loaded defaults from res/ezycad_settings.json");
   return result;
 #endif
-}
-
-std::string load_with_defaults()
-{
-  std::string content = load();
-  if (content.empty())
-  {
-    content = load_defaults();
-    if (!content.empty())
-      save(content);
-  }
-  return content;
 }
 
 void save(const std::string& content)
